requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
distlib==0.3.7
dnspython==2.4.2
filelock==3.12.3
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
nltk==3.8.1
numpy==1.25.2
pandas==2.1.0
platformdirs==3.10.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.8.8
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
virtualenv==20.24.4
Werkzeug==2.3.7
